# ticketing-system/openapi.yaml
openapi: 3.0.3
info:
  title: Ticketing System API
  version: 0.2.0
  description: |
    Jira-like ticketing system API with auth, Kanban workflow, projects, groups, and webhooks.
servers:
  - url: http://localhost:8080
...
security:
  - sessionAuth: []
paths:
...
  /projects/{projectId}/stories:
    get:
      summary: List stories
      operationId: listStories
      tags: [tickets]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Story list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryListResponse"
    post:
      summary: Create story
      operationId: createStory
      tags: [tickets]
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryCreateRequest"
      responses:
        "201":
          description: Story created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
  /stories/{id}:
    get:
      summary: Get story
      operationId: getStory
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
    patch:
      summary: Update story
      operationId: updateStory
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryUpdateRequest"
      responses:
        "200":
          description: Story updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
    delete:
      summary: Delete story
      operationId: deleteStory
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
  /tickets/{id}/comments:
    get:
      summary: List ticket comments
      operationId: listTicketComments
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Comment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketCommentListResponse"
    post:
      summary: Add ticket comment
      operationId: addTicketComment
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketCommentCreateRequest"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketComment"
  /tickets/{id}/comments/{commentId}:
    delete:
      summary: Delete ticket comment
      operationId: deleteTicketComment
      tags: [tickets]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: commentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: ticketing_session
  schemas:
...
    Story:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, projectId, title, createdAt, updatedAt]
    StoryCreateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      required: [title]
    StoryUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
    StoryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Story"
      required: [items]
    TicketComment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ticketId:
          type: string
          format: uuid
        authorId:
          type: string
          format: uuid
        authorName:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, ticketId, authorId, authorName, message, createdAt]
    TicketCommentCreateRequest:
      type: object
      properties:
        message:
          type: string
      required: [message]
    TicketCommentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TicketComment"
      required: [items]
...